fix #557 blur agenda item input after item gets added
